backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.1
djongo==1.3.6
dnspython==2.4.2
email-validator==2.1.0
python-decouple==3.8
requests==2.31.0
pytest==7.4.3
pytest-django==4.7.0
